renamed image and alts to be SHED not the SHED
