+++ client/src/pages/lab_management/storefront_preview/NoTransactionView.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Stack, Typography } from "@mui/material";
-import LogoSvg from "../../../assets/the shed logo orange white.png";
+import LogoSvg from "../../../assets/shed logo orange white.png";
 import styled from "styled-components";
 import NameTag from "./NameTag";
